# Use lazy_static to parse language syntax once