--- a/Cargo.toml
+++ b/Cargo.toml
@@ -15,6 +15,7 @@ piston_meta = "0.26.1"
 range = "0.3.1"
 rand = "0.3.13"
 read_color = "0.1.0"
+lazy_static = "0.2.1"
 
 [dependencies.hyper]
 version = "0.9.7"