@bvssvni 2016-09-06
[Tested](https://github.com/PistonDevelopers/dyon/issues/52) n_body benchmark with 1 million steps (copied code instead of writing loader): Rust is 5.5x faster in debug mode, 191.4x faster in release mode.