@bvssvni 2016-09-06
Dyon vs Python vs Javascript on the iterate benchmark: Dyon is 2.1x faster than Python, Javascript is 11.2x faster than Dyon.